finalize component structure
